Updates to the cache and update handling for name_db.py.
These updates are a response to two technical issues:

1. The database connection latency is causing the load time to be
too slow (minutes). A cache system for the document data was
added to reduce the queries required for loading.
This includes a process to preload the entire existing db.

2. SIMBAD is giving main_ids values that have different
capitalization compared to the main_id for an existing
record for the same star.
We needed to add case-insensitive id checking to
and to be certain that updates to the original record
preserve the exact (capitalization) string of the original _id.
This is required for the MongoDB database when using
the SIMBAD MAIN_ID as the equivalent _id field in MongoDB.