Created a new Bson validation types for the website's database.
This is the start of a process to have the components for the website database
build at the same processing step in the NatCat python class.

As a part of this effort, we are moving some dictionaries and lists
that were previously hard coded for the website
to new homes in a separate directory with the
other data and version-controlled configurations.