Updated stellar parameter handing to enforce units and parameter types.
Updated Gaia Reference data is including a parallax distance.
That calculation is now reformed only when necessary for a limited
number of stars, ~109.

Changed the output of all the stellar parameter data
formatted in a common way before being included in a database.
This allowed code to be deleted that did this correction
at a later point with other calculations,
and was a sink of energy doing a very confusion set of operations.