fixed - absolute abundances being excluded during un-normalization.
Now, absolute abundances are not required to have a corresponding values in that catalog's normalization key.
This requirement was excluding abundances that were already 'absolute' and did not need an un-normalization,
if those the element was not in the un-normalization references for that catalog,
which was an unnecessary extra step.