MongoDB admin tools.
Added an argparse interfaces MongoDB administrative tasks.
This includes the ability to add and delete users.
When users are added, an `.env` file is created in the
format expected by the HySite repo.

URL encoding to the MongoDB connection_string allows us
to use passwords with special characters.
Random 100 character passwords are out default.