Read-only mode
Added a new feature to allow some capabilities of the
HySite repository to be used in a read-only mode that can
fetch data from the hypatiacatalog.com, but is not able
to post processed data.

This mode lacks raw import files.
Currently, those trigger warnings, but in the future they
might instead simply fetch the required data from the Hypatia API.
But this will require new endpoints and pipeline processes.

The only demonstration of this mode is all_data.py,
a new example file.