Fix for issue #5
Multiple fixes that were required for writing catalog files.
File writing issues caused the uniquify processing for elements,
to fail when as the written files contained many mistakes.

After fixing the writing errors, these files needed to be read in correctly.
This process also had to be updated, as the simbad names and wording processing
can skip the name parsing scripts that require us to update parsing functions.
The ongoing removal of the original the star name parsing tools.