Gaia data migrated to MongoDB.
Updates to the Hypatia pipeline.
Moved `convert_to_object_params` to its own `def`
and not a static-method of GaiaLib.

Minor interface updates to the GaiaLib in sources/gaia/ops.py,
as the database interface was simplified from this update.

Testing allowed for a successful upload.
Merging into the main branch.