New update processes for the pipeline to the live website.
We have removed the language update process from
backend/update.py and moved it to backend/hypatia/legacy/update.py
This allows use to use these files to make updates but without importing
pandas, hdf5, and other legacy packages that are no longer needed.

The read-only files system for docker containers is now toggleable
(see example.env) but is false by default, but true on the server.
This allows us to use the docker environment instead of a
python environment which continues to make a local python
installation optional.

The publication scripts were moved into the update.py.
It is toggled with the --publish command.
This was done to make fewer interfaces for a user during
the update process.

Big updates to the README.md file to include all the examples
and usecases that are need for installation and updates.
This should be split into a wiki in a future upgrade
to improve readability and aid in navigation.