Updated definition to create output cross-matched to a target list.
There is a definition in standard_lib.py that was set up to output Hypatia data in 2D (create_flat_file). It's now been updated to accept a target list and retain the original star name within the output file, which will be in the same order as the target list.

Also updated a solar normalization based on an email from the author.