Add SIMBAD parameters to be included in the database uploaded.
Simplified the simbad query proces to only use the tap queries,
as a result tap.py was deleted.

The validator.py was created to house the validator for db.py.

This includes the spectral type and reference
and also the reference for the ra, dec hms.

Move the spectrum parser to elements.py, in prep for deleting
legacy data processing tools.