Moved star_count_per_element_histogram() to use the live site api endpoint and not a pipeline calculation.
Using MongoDB and the live website is much faster
than running the pipeline and using is inspection tools.
This ensures that the data used is the same as what is
deployed on the live website and is not effected by
changes from the various sources that contribute
to the hypatia-catalog pipeline.