Configuration data refactor, files added to backend/hypatia/config directory.
config.py file was split into three new files not in backend/hypatia/config,
1. file_paths.py
2. env_load.py
3. source_settings.py

This was meant to make the look-up process simply for these setting.

params_units.toml was removed from HyData and added to
backend/hypatia/config/params_units.toml.
This choice was made to move separated science data sources
from configurations and settings that do not require reference citations.
It should also make it easier for external collaborators to use
the HySite repo with other the source abundance catalogs.

element_data.py was moved to backend/hypatia/config

The output data directory is now meant to be initialized when
the HySite repository is cloned from git. This was done with
hierarchy of .gitignore files, a typical trick in this repository.
This is to control the permissions for directories more explicitly
for Docker and python users without the permission to make directories.