Missing data file handling with a database query, then hypatiacatalog.com API fetch.
This is more on removing the HyData dependency.
HyData is needed for data pipeline operations,
but is not needed on the Webserver or for local website/API/DB operations.

Now when the data files for solar-normalization and representative error are not found,
this data is fetched from the database using only a read connection,
and the target does not exist, it downloads it from hypatiacatalog.com.
If that fails, the processes simply display a warning and continue.

Representative error was split from the elements.py and put into
the new element_error.py.

Dependencies were moved out of summary.py to avoid circular imports.