Added hover text to stellar and planet params (issue $16); other table formatting and bug fixes
Changed the default table length to 1000 items.

Changed "load 100 more rows" bottom to simply load all the available rows of data.

Fixed a bug formatting all numbers to two decimal places.
This now comes from the params_units.toml that is loaded to the API.

Fixed a bug with `period` units not being shown on the table header.