Created duplicate catalog_file for testing and added attention messages.
Data for an upcoming paper needed to be included in the Hypatia machinery to view supplemental data and do comparisons with literature. However, this dats is not ready for the website. So, a new catalog_file_NAME file was created (in .gitignore) and the file_path of the default catalog_file was changed. To make sure that I don't forget about this change, I included yellow attention text in the new_catalog file as well as the output.py file so the file version is obvious.