Bug fixes for element names.
The variable name for elements was simplified in the pipeline (Fe), but the filtering was using the older more complex projection (normalization.lodder09.Fe.value). This was causing the element filtering to fail. This is in aggregation.py.

Added new testing points in data_process.py and db.py to allow use of the debugger.

Added more pipeline stages to be printed to the screen in DEBUG mode.