Cleaned up redundant code.
Removed the old multi_output from standard_lib and also got rid of the mdwarf_output in sandbox. Needed a place to put the mdwarf_histogram, so put it back in sandbox for now (couldn't have it in sakhmet_tools because of circular imports) but will move it to a better location later.