Updated to display a hybrid web2py and NextJS frontend application.
This required getting the application to build in Docker, and fetch data during the build process.
This is complicated enough that we introduced the script display.sh

Add the frontend/components directory to hold reusable application code that is not specific to a page. This includes header.tsx and image.tsx.

Added page internal page routing definitions to api_targets.tsx.